# core: speed up checking for uses