Before:

```
❯ python benchmarks/rewriting.py RemoveUnused.is_trivially_dead dis
===========================================
 Events recorded: 58
 Opcodes recorded: 40
 Uninstrumented: 350.35 ns
 Uninstrumented per opcode: 8.76 ns/op
 Instrumented: 11895.62 ns
 Instrumented per opcode: 297.39 ns/op
 Elapsed: 4825.74 ns
 Elapsed per opcode: 120.64 ns/op
 Opcode overhead: 0.00 ns/op
 Tracing overhead per opcode: 111.88 ns/op
===========================================
```

After:

```
❯ python benchmarks/rewriting.py RemoveUnused.is_trivially_dead dis ===========================================
 Events recorded: 39
 Opcodes recorded: 27
 Uninstrumented: 220.84 ns
 Uninstrumented per opcode: 8.18 ns/op
 Instrumented: 7753.91 ns
 Instrumented per opcode: 287.18 ns/op
 Elapsed: 2951.75 ns
 Elapsed per opcode: 109.32 ns/op
 Opcode overhead: 0.00 ns/op
 Tracing overhead per opcode: 101.15 ns/op
===========================================
```