@superlopuh 2025-08-01
I agree, but I also think that it's worth it for this code, I'll merge it as-is.